# order/src/main/resources/api.yaml
openapi: 3.0.0
servers:
  - description: 摇一摇服务端
    url: 'http://192.168.10.156:8080/'
info:
  description: |
    摇一摇
  version: 1.0.2
  title: XWalletServer
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: liujian@xender.com
tags:
  - name: shake
    description: 摇一摇发现文件
  
paths:
  /order/upload:
    post:
      tags:
        - shake
      summary: 上传视频缩略图和视频文件信息
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
        default:
          $ref: '#/components/responses/ApiError'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                info:
                  $ref: '#/components/schemas/VideoInfo'
  /order/files:
    get:
      tags:
        - shake
      summary: 根据用户ID查询文件
      parameters: 
      - name: userId
        in: query
        description: 用户ID
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoInfo'
            
        default:
          $ref: '#/components/responses/ApiError'
    
     

externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
components:
  schemas:
    VideoInfo:
      type: object
      properties:
        id: 
          type: integer
          format: int64
        userId:
          type: string
        path:
          type: string
        name:
          type: string
        size:
          type: integer
          format: int64
        thumbnailPath:
          type: string
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: integer
          format: int64
  responses:
    ApiError:
      description: operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'        
        
        
        
          
          
          
          
          
          
          
          
 